requirements.txt
streamlit
torch
Pillow
torchvision
transformers
numpy
boto3
nltk
pycocotools
ultralytics